refactored components outside of one file and into components folder
