+++ client/src/components/HomePage/constants.ts
@@ -1,0 +1,43 @@
+export const RESOURCE_LINKS = [
+  {
+    text: "Career & Talent Development",
+    url: "https://career.fiu.edu/",
+  },
+  {
+    text: "FIU Academic Calendar",
+    url: "https://onestop.fiu.edu/academic-calendar/",
+  },
+  {
+    text: "CASE Drop-In Advising",
+    url: "https://case.fiu.edu/advising/",
+  },
+  {
+    text: "Undergraduate Advising",
+    url: "https://www.cis.fiu.edu/academics/advising/undergraduate/",
+  },
+  {
+    text: "Graduate Advising",
+    url: "https://cec.fiu.edu/students/academic-advising/graduate-advising/",
+  },
+];
+
+export const FEATURES = [
+  {
+    step: "1",
+    title: "Upload Your Audit",
+    description:
+      "Upload your degree audit and we'll analyze your academic progress instantly.",
+  },
+  {
+    step: "2",
+    title: "Ask Anything",
+    description:
+      "Ask about courses, prerequisites, graduation requirements, or academic policies.",
+  },
+  {
+    step: "3",
+    title: "Get Personalized Advice",
+    description:
+      "Receive tailored recommendations based on your specific academic situation.",
+  },
+];
